+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,7 +5,7 @@
 import FlutterMacOS
 import Foundation
 
-import icon_switcher
+import basic_icon_switcher
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   IconSwitcherPlugin.register(with: registry.registrar(forPlugin: "IconSwitcherPlugin"))