+++ ios/Classes/SwiftIconSwitcherPlugin.swift
@@ -4,7 +4,7 @@ import UIKit
 public class SwiftIconSwitcherPlugin: NSObject, FlutterPlugin {
 
     public static func register(with registrar: FlutterPluginRegistrar) {
-        let channel = FlutterMethodChannel(name: "icon_switcher", binaryMessenger: registrar.messenger())
+        let channel = FlutterMethodChannel(name: "basic_icon_switcher", binaryMessenger: registrar.messenger())
         let instance = SwiftIconSwitcherPlugin()
         registrar.addMethodCallDelegate(instance, channel: channel)
     }