+++ windows/basic_icon_switcher_plugin_c_api.cpp
@@ -1,12 +1,12 @@
-#include "include/icon_switcher/icon_switcher_plugin_c_api.h"
+#include "include/basic_icon_switcher/basic_icon_switcher_plugin_c_api.h"
 
 #include <flutter/plugin_registrar_windows.h>
 
-#include "icon_switcher_plugin.h"
+#include "basic_icon_switcher_plugin.h"
 
 void IconSwitcherPluginCApiRegisterWithRegistrar(
     FlutterDesktopPluginRegistrarRef registrar) {
-  icon_switcher::IconSwitcherPlugin::RegisterWithRegistrar(
+  basic_icon_switcher::IconSwitcherPlugin::RegisterWithRegistrar(
       flutter::PluginRegistrarManager::GetInstance()
           ->GetRegistrar<flutter::PluginRegistrarWindows>(registrar));
 }